# docs/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"AI Services API","description":"API reactiva para generación de imágenes (Text-to-Image) y eliminación de fondos (Background Remover).","contact":{"name":"Andre Hermoza","email":"andre.hermoza@example.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Servidor configurado"}],"tags":[{"name":"Background Remover","description":"Eliminación de fondos en imágenes mediante IA"},{"name":"Text to Image","description":"Generación de imágenes con IA a partir de prompts de texto"}],"paths":{"/api/v1/tti/generate":{"post":{"tags":["Text to Image"],"summary":"Generar imagen","description":"Envía un prompt a la IA y persiste el resultado","operationId":"generate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateImageRequest"}}},"required":true},"responses":{"200":{"description":"Imagen generada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/textToImage"}}}}}}},"/api/v1/bgremover/process":{"post":{"tags":["Background Remover"],"summary":"Procesar imagen","description":"Recibe un archivo de imagen, elimina el fondo y persiste el resultado","operationId":"process","requestBody":{"description":"Archivo de imagen a procesar","content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}},"required":true},"responses":{"200":{"description":"Imagen procesada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BGRemover"}}}}}}},"/api/v1/tti/deactivate/{id}":{"patch":{"tags":["Text to Image"],"summary":"Desactivar registro","description":"Cambia el estado del registro a inactivo","operationId":"deactivate","parameters":[{"name":"id","in":"path","description":"Identificador de la generación","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/textToImage"}}}}}}},"/api/v1/tti/activate/{id}":{"patch":{"tags":["Text to Image"],"summary":"Activar registro","description":"Cambia el estado del registro a activo","operationId":"activate","parameters":[{"name":"id","in":"path","description":"Identificador de la generación","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/textToImage"}}}}}}},"/api/v1/bgremover/deactivate/{id}":{"patch":{"tags":["Background Remover"],"summary":"Desactivar registro","description":"Cambia el estado del registro a inactivo","operationId":"deactivate_1","parameters":[{"name":"id","in":"path","description":"Identificador del procesamiento","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BGRemover"}}}}}}},"/api/v1/bgremover/activate/{id}":{"patch":{"tags":["Background Remover"],"summary":"Activar registro","description":"Cambia el estado del registro a activo","operationId":"activate_1","parameters":[{"name":"id","in":"path","description":"Identificador del procesamiento","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BGRemover"}}}}}}},"/api/v1/tti":{"get":{"tags":["Text to Image"],"summary":"Listar generaciones","description":"Obtiene el historial de imágenes generadas","operationId":"findAll","responses":{"200":{"description":"Listado obtenido correctamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/textToImage"}}}}}}}},"/api/v1/tti/{id}":{"get":{"tags":["Text to Image"],"summary":"Obtener por ID","description":"Busca una generación por su identificador","operationId":"findById","parameters":[{"name":"id","in":"path","description":"Identificador de la generación","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Generación encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/textToImage"}}}}}}},"/api/v1/bgremover":{"get":{"tags":["Background Remover"],"summary":"Listar procesamientos","description":"Obtiene el historial de imágenes procesadas","operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BGRemover"}}}}}}}},"/api/v1/bgremover/{id}":{"get":{"tags":["Background Remover"],"summary":"Obtener por ID","description":"Busca un procesamiento por su identificador","operationId":"findById_1","parameters":[{"name":"id","in":"path","description":"Identificador del procesamiento","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BGRemover"}}}}}}},"/api/v1/bgremover/images/{filename}":{"get":{"tags":["Background Remover"],"summary":"Descargar imagen procesada","description":"Sirve una imagen PNG almacenada localmente","operationId":"getImage","parameters":[{"name":"filename","in":"path","description":"Nombre del archivo de imagen","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Imagen encontrada","content":{"image/png":{}}}}}}},"components":{"schemas":{"GenerateImageRequest":{"type":"object","description":"Solicitud para generar una imagen a partir de un prompt de texto","properties":{"prompt":{"type":"string","description":"Descripción textual de la imagen a generar","example":"a futuristic city at sunset"}}},"textToImage":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"prompt":{"type":"string"},"generated_image_url":{"type":"string"},"status":{"type":"boolean"},"created_at":{"type":"string","format":"date-time"}}},"BGRemover":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"source_image_url":{"type":"string"},"processed_image_url":{"type":"string"},"status":{"type":"boolean"},"created_at":{"type":"string","format":"date-time"}}}}}}